Split chainseekd in two
